Updated logic of listing real network interfaces
The script will now look for network devices inside /sys/class/net
that have a folder called `device` (a Network Interface Card) instead
of doing string operations in `ip link show`.

It is unpredictable and the previous logic only checks if an interface
has a MAC address for it to be real, but virtual interfaces can also
have a MAC address likeVirtualbox Host Only Ethernet Adapter (vboxnet0)
and docker networks.